Restructure Category and Entry so both have one component and are grouped in the same folder (categoryCreateOnRegister as an exception)
