--- budget-binder-multiplatform-app/src/jsMain/kotlin/de/hsfl/budgetBinder/compose/categoryCreate/CategoryCreateComponent.kt
@@ -1,28 +1,0 @@
-package de.hsfl.budgetBinder.compose.categoryCreate
-
-import androidx.compose.runtime.*
-import de.hsfl.budgetBinder.domain.use_case.auth_user.LogoutUseCase
-import de.hsfl.budgetBinder.domain.use_case.get_user.UserUseCase
-import de.hsfl.budgetBinder.presentation.LogoutViewModel
-import de.hsfl.budgetBinder.presentation.Screen
-import de.hsfl.budgetBinder.presentation.UiState
-import de.hsfl.budgetBinder.presentation.UserViewModel
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.SupervisorJob
-import org.kodein.di.compose.localDI
-import org.kodein.di.instance
-
-@Composable
-fun CategoryCreateComponent(screenState: MutableState<Screen>) {
-    val scope = CoroutineScope(Dispatchers.Unconfined + SupervisorJob())
-    val di = localDI()
-    val userUseCase: UserUseCase by di.instance()
-    val userViewModel = UserViewModel(userUseCase, scope)
-    val viewState = userViewModel.state.collectAsState(scope)
-
-    CategoryCreateView(
-        state = viewState,
-        onBackButton = { screenState.value = Screen.CategorySummary} //Should go back to the previous Screen, which could be CategorySummary or EntryCreate.
-    )
-}
